deleteImageを切り出した
